scientific_parser/requirements.txt
pydantic-ai 
mcp 
a2a-python
pydantic-graph
arxiv
google-generativeai
python-dotenv
requests
rich
asyncclick
fasta2a
flask
click